[BUGFIX] tca_media_field.php refers to mmforeign
There are some remains of "mmforeign" in tca_media_field.php.

Solution:
As the userProcessClass is unset anyway if the TYPO3 version is higher
than 4.1, it can be safely removed.

At the same time, the "MM_foreign_select" value can be removed as it is
marked as obsolete since 4.1.

Change-Id: I599610194c447a85bc79f10c4b40ab9c18bd5336
Resolves: #M14927
Releases: 1.2