[BUGFIX] Media type detection not working properly
In http://bugs.typo3.org/view.php?id=10556 the internal
DAM arrays to match a file extension to a media code were
removed to prioritize the information from the table
tx_dam_media_types. Because of an error this didn't work
properly leading to an incorrect media type assignment
(e.g. pdf and eps => application).

This patch radically removes the remains of the internal arrays
and makes DAM depend only on the table mentioned above.

Change-Id: I5fdfd96a70422e370aeacb22293e3b139624cd82
Resolves: #M18090
Releases: 1.2