[BUGFIX] Thumbs in asset edit: Do not scale up small images
If you have very small images (e.g. icons) they
get scaled up in thumbs view. If the images are
really tiny, they are hardly recognizable.

With this change images smaller than the thumbnail
container are not scaled up anymore but kept on
original size and centered in the container.

Change-Id: I70e9aaf84d8eaef50d3022d4b33adc1710c96523
Resolves: #M8647
Releases: 1.2