[BUGFIX] Search string in list lost when using pagination
If the DAM list is filtered by a search term and the pagination is
used, the search term is lost on the target page. This is because
the search term is not passed to the target page.

This change checks if a search term is set. If that's the case, it's
added to the link parameters of the previous/next links/buttons in
the list module.

While investigating, I found out that the previous/next links
on the top and bottom of the list (e.g. [21-40]) display wrong numbers.
This is also fixed.

Change-Id: I2052a30b37847a04a936c3ae5ebd0a7b608d655c
Resolves: #M12952
Releases: 1.2