[BUGFIX] Missing icon in file delete dialogue
The page icon is missing in the list of usages for a certain DAM asset
when you try to delete it.

t3lib_iconWorks::getIconImage doesn't seem to be working anymore for
generating an icon for the table "pages". I replaced the call to
getIconImage by a call to the Sprite Icon API.

No-brainer.

Change-Id: I8e49c3d5190885078ee748c87d3a18be06b1207c
Resolves: #M18318
Releases: 1.3