[BUGFIX] All translated deleted when checking the index
When you use the "Check index" function in the "Tools" module and you
check the "Mark entries with missing files as deleted" box, all
translation DB entries will also be marked as deleted.

The language overlay doesn't hold the path and name of the file,
therefore the file is not found and the language overlay is deleted.

This change fixes these it by doing the following:
- DAM does not check anymore for file existence of records with
sys_language_uid other than 0
- If a DAM record needs to be deleted because a file is missing,
its language overlays is be deleted as well

Change-Id: I16a9d7d2ebe5c5748dadd0249c4700093a86fc17
Resolves: #M18402
Releases: 1.2