[BUGFIX] Files with comma in filename break batch-edit mode
The selected files are passed to the script in a checkbox array.
On the way to become parameters of the folder the user selects for
moving/copying the file to this array is imploded to a string with
comma as a delimiter. Later it is exploded and so if a filename
has got commas, the references are wrong.

So the delimitier was changed to ###ITEMSEPARATOR###.

Change-Id: I9a08761da17306be5502120a4faf788b277bb95e
Resolves: #M11331
Releases: 1.2