[TASK] Cleanup TCA of tx_dam
The TCA of tx_dam has an extensive first tab. Similar to the TCA in
pages or tt_content, the access information should better be in a
separate tab.

At the same time the category selector should have a bigger default
size (10 instead of 6).

Due to missing styles in T3SKIN, some palette headers were not indented
correctly. This is fixed as well.

Change-Id: I6308ced56f54f421a9360ed3c952aab13fd0dff4
Resolves: #M6919
Releases: 1.3