[BUGFIX] Fatal error when clicking on info icon in Media > File
With DAM and TYPO3 4.7.0, clicking on the info icon of a file in the
Media > File module in list mode opens a popup that shows the following
error message:

Fatal error: Call to undefined method template::getDynTabMenuJScode()

This method is deprecated since TYPO3 4.5 and is no longer necessary
since the DynTabMenu JavaScript code is inserted automatically if there
is one. Therefore the call can be removed without replacement.

Change-Id: I35103a275cd93861c6413f00994f41db590a92c4
Resolves: #M18526
Releases: 1.3