[FEATURE] Usage information items should display page path
In the usage information tab of an item the name of the page where
the item is used is displayed. If you have may pages with the same
page title it's hard to find out on which page the item is
referenced.

Therefore the page path needs to be displayed on hovering the page
title.

Change-Id: If01ef1f1f5a418974296fbfed80b2b820b8bb940
Resolves: #M10233
Releases: 1.3