[FEATURE] DAM references for Media CE
This adds a new media wizard provider for DAM elements. The path to the
audio or video file must have the syntax "media:[UID]".

The changes in http://bugs.typo3.org/view.php?id=14893 provide a media
tab for the normal TYPO3 link wizard to reference DAM records in that
syntax.

Change-Id: I2248d2a4310987fe98b36134aff3b307ea65e05a
Resolves: #M11909
Releases: 1.3