[TASK] Optional disabling of workspaces for DAM records
As discussed with the workspaces team, versioning for DAM records and
DAM category records should not be versioned anymore because there
are a lot of problems with workspaces and DAM.

This change introduces a switch in the extension configuration to
turn versioning off, thus it remains backwards compatible with
installation that are actually using DAM and workspaces.

At the same time the old, no longer working "versioning" configuration
is removed from tx_dam_selection and tx_dam_media_types.

Change-Id: Ib57bbd5ee0f0e7738501c01d50a5b3088dea98fb
Releases: 1.3
Fixes: #M18405