[BUGFIX] Wrong categories upon translation
When a Media (tx_dam) element is translated, the translation
is assigned wrong categories in the tx_dam_mm_cat table.
In fact it should not be assigned any categories at all, since
it is just a translation. This can be fixed by really excluding
the field from translation by dropping usage of "l10n_display".

Change-Id: I4822c73df266e272f456e9c92bcc0dc2eac2799a
Fixes: #M18414
Releases: 1.3