[BUGFIX] keepFileName does not respect change of file extension
When the indexing option keepFileName is set, the file name of a
file is kept on replacing. This only makes sense if the file type
of the new file is the same file type as the old one.

Therefore the setting keepFileName is ignored when the file types
don't match.

Change-Id: Id36e1fb1fa987209b182602f6ea8218088f8b90a
Relates: #M13455
Resolves: #M18595
Releases: 1.3