[BUGFIX] Regression: No file uploaded notice with empty file inputs
This is a regression from #M15809. In #M15809 a post-upload check was
implemented to find out whether a file was sent with the POST request
or not. This leads to notices that "No file was uploaded" when a user
uploads only one file in a form of more than one file.

This is fixed with this changeset and the error message is outsourced
to the language file.

Change-Id: Id619030c519c351b71d9b9229b89fa95d99cc2ae
Relates: #M15809
Fixes: #M18566
Releases: 1.3