[BUGFIX] Media link handler doesn't respect Typolink configuration
The media link handler that enables TYPO3 to set header links and image
links to DAM items, drops the Typolink configuration before generating
the link to the file.

This makes e.g. the jumpurl feature unusable. In fact only the default
parameter must be unset and replaced by the DAM item uid.

Change-Id: I98ce0cd1f9e9e85748549fe384f9438f1a9bc37f
Fixes: #M18617
Release: 1.3