[!][BUGFIX] Link title attribute from RTE cannot be overridden
DAM automatically sets title attributes for links set in RTE. The fields
taken to generate this attribute can be configured in the RTE
configuration.

Caurrently, a title tag set in RTE, cannot be overriden for frontend
rendering. The title configuration of the media tag
(plugin.tx_tsfemediatag) is only considered when there is no title set in
the RTE.

The title from the RTE is now set to
plugin.tx_dam_tsfemediatag.tag.typolink.title.dataWrap

Therefore it can be overriden in any case:
dataWrap.override = {field:txdam_file_name}

A fallback can be defined:
dataWrap.ifEmpty = {field:txdam_file_name}

This is only a breaking change if you use a custom TS configuration for the
media tag. Then you need to rename dataWrap to dataWrap.ifEmpty to restore
the previous behaviour.

Change-Id: Ic286adae2d65b26482bc26358f5f38c59e4c9cc0
Fixes: #M18641
Release: 1.3