[BUGFIX] Media > File has no default sorting
Since #M18573 the configuration for DAM modules is taken from the
user configuration, this includes the information about sorting
columns.

As long as the user didn't set a sorting column, the entries are mixed
up. Therefore a default sorting column (folder name/file name) must be
added to the extension.

Change-Id: I2b1706792ed98a380f6fbe53590c05c8ed3342d6
Fixes: #M18608
Release: 1.3