[BUGFIX] Auto indexing must be disabled in FE mode
Auto indexing is enabled by default in DAM. Due to a bug this setting
was not respected until DAM 1.3.1. This led to another bug: DAM is
trying to index in FE mode. But without a logged in backend user
writing to the tables and logs is not possible.

Therefore auto indexing must be disabled in frontend mode.

Change-Id: I5df74e45749859b72f170053ff4a5fafe07f53b4
Release: 1.3
Fixes: #41619