[BUGFIX] Log messages for move and delete miss file information
When DAM log messages contain data about files with long names or paths,
those data do not get fully stored in the sys_log table because the
log_data field is limited to 255 characters.

This patch removes a part of the file paths, namely PATH_site.
With this shorter paths, there is a higher chance of the log data being
stored in full length.

Change-Id: Ie293572b4c14f3092bfbd1df06926ed9b9ba11e1
Fixes: #42418