[TASK] On replacing a file, cache of the pages using it should be cleared
The the DAM backend module you can replace a file with another file.
All links to the file stay intact.

If the file name of the new file differs from the original file name,
you'll run into the problem that the cache is not cleared after
replacing. The old file doesn't exist anymore, but the page linking to
it is still cached and therefore you will have a dead link until the
cache of the page is rebuilt.

The solution is to clear the page of affected pages after replacing the
file.

Change-Id: I0a9c7bc824ef3a3d7358ff1b9cebe51de61d1b13
Release: 1.3
Fixes: #43149
Reviewed-on: https://review.typo3.org/16603
Reviewed-by: Stefan Neufeind
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich