[BUGFIX] Respect TCAdefaults
In tx_dam_db::insertUpdateData(), some code exists to read TCAdefaults
set per User TSconfig, but it is not used. Indeed one more call to
TCEmain must be made to get an "empty" data array prefilled with TCA
default override values. This data must be retrieved and merged with
the Page TSconfig and the meta data.

Furthermore this operation should be performed be performed only
for a new record (test is currently incomplete).

Resolves: #43308
Releases: 1.3

Change-Id: I12822b542353d76c3917ac8c8d6a28aa9d6d56c9
Reviewed-on: https://review.typo3.org/16741
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich