[BUGFIX] Replacing a file can destroy the file of another DAM record
Replacing a file with a file whose name is identical to the file name of
another record leads to database inconsistency because the file of the
other record is overwritten.

Example:
File1.jpg -> Title "Example"
File2.jpg -> Title "Example2"

If you choose "Replace File" on Example2 and upload a file with the
filename "File1.jpg" you'll end up with two entrys that reference to the
same (new) file.
File1.jpg -> Title "Example2"
File1.jpg -> Title "Example2"

This is because DAM doesn't check if the file to be replaced is
overwritten or a file from any other record that by chance has the
same file name.

Change-Id: Ic07f48e72553f104e026c58e33674387645bd50c
Releases: 1.3
Fixes: #42040
Reviewed-on: https://review.typo3.org/18381
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich