[BUGFIX] DAM breaks Element Browser tab change
When the Element Browser is called from a TCEform field (such as
header_link/image_link), switching from the Media tab to another
tab makes the Element Browser lose context.

This is because the function jumpToUrl was overwritten from a
DAM function with the same name that is used in the backend modules.
An additional check for the context needs to be added to prevent this
error.

Additionally, the JavaScript code for the context menu is removed from
browse_media class to prevent a JavaScript error.

Change-Id: Ie2d1a7b1710f605e84383e2538bd0fd1d7e9a74b
Resolves: #43714
Releases: 1.3
Reviewed-on: https://review.typo3.org/18550
Tested-by: Karsten Madsen
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske